requirements.txt
aws-cdk-lib==2.102.0
constructs>=10.0.0,<11.0.0
pydantic==2.4.2
PyYAML==6.0.1
Jinja2==3.1.2
python-dotenv==1.0.0
mypy==1.6.1